figuring out where blocks are

+++ implementations/cpu_pathfinder_array.py
@@ -214,7 +214,7 @@ def search(grid, start, goal, open, closed, parents, cost, g, h, UNEXPLORED, nei
         getNeighbors(grid, current, neighbors, direction)
         for next in neighbors:
         # for next in getNeighbors(grid, current, neighbors):
-            next_x = int((next-(next%width))/width)
+            next_x = (next-(next%width))/width
             next_y = next%width
             print(next_x, next_y)
             if passable(grid, (next_x, next_y)) and inBounds(grid, (next_x, next_y)):
